+++ libs/types/src/epoch.rs
@@ -1,0 +1,53 @@
+// Copyright 2021 Centrifuge Foundation (centrifuge.io).
+//
+// This file is part of the Centrifuge chain project.
+// Centrifuge is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version (see http://www.gnu.org/licenses).
+// Centrifuge is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+use cfg_primitives::Moment;
+use frame_support::traits::Get;
+use codec::{Decode, Encode, MaxEncodedLen};
+use frame_support::{traits::UnixTime, RuntimeDebug};
+use scale_info::{build::Fields, Path, Type, TypeInfo};
+
+use crate::{pools::PoolChanges, tranches::EpochExecutionTranches};
+
+#[derive(Clone, Encode, Decode, Eq, PartialEq, RuntimeDebug, TypeInfo)]
+pub struct EpochState<EpochId> {
+	/// Current epoch that is ongoing.
+	pub current: EpochId,
+	/// Time when the last epoch was closed.
+	pub last_closed: Moment,
+	/// Last epoch that was executed.
+	pub last_executed: EpochId,
+}
+
+#[derive(Encode, Decode, Clone, Eq, PartialEq, RuntimeDebug, TypeInfo)]
+pub struct ScheduledUpdateDetails<Rate, MaxTokenNameLength, MaxTokenSymbolLength, MaxTranches>
+where
+	MaxTokenNameLength: Get<u32>,
+	MaxTokenSymbolLength: Get<u32>,
+	MaxTranches: Get<u32>,
+{
+	pub changes: PoolChanges<Rate, MaxTokenNameLength, MaxTokenSymbolLength, MaxTranches>,
+	pub scheduled_time: Moment,
+}
+
+/// The information for a currently executing epoch
+#[derive(Encode, Decode, Clone, Eq, PartialEq, RuntimeDebug, TypeInfo)]
+pub struct EpochExecutionInfo<Balance, BalanceRatio, EpochId, Weight, BlockNumber, TrancheCurrency>
+{
+	epoch: EpochId,
+	nav: Balance,
+	reserve: Balance,
+	max_reserve: Balance,
+	tranches: EpochExecutionTranches<Balance, BalanceRatio, Weight, TrancheCurrency>,
+	best_submission: Option<EpochSolution<Balance>>,
+	challenge_period_end: Option<BlockNumber>,
+}
