+++ libs/types/src/lib.rs
@@ -16,6 +16,7 @@
 
 ///! Common-types of the Centrifuge chain.
 pub mod adjustments;
+pub mod epoch;
 pub mod fee_keys;
 pub mod fixed_point;
 pub mod fulfillments;
