+++ libs/types/src/reserves.rs
@@ -13,15 +13,20 @@
 use cfg_primitives::types::Balance;
 use cfg_traits::InvestmentProperties;
 use codec::{Decode, Encode, MaxEncodedLen};
-use frame_support::{traits::UnixTime, RuntimeDebug};
+use frame_support::{dispatch::DispatchResult, traits::UnixTime, RuntimeDebug};
 use scale_info::{build::Fields, Path, Type, TypeInfo};
 #[cfg(feature = "std")]
 use serde::{Deserialize, Serialize};
-use sp_runtime::{traits::Zero, Perquintill};
+use sp_runtime::{
+	traits::{AtLeast32BitUnsigned, Zero},
+	ArithmeticError, Perquintill,
+};
 use sp_std::{
 	cmp::{Ord, PartialEq, PartialOrd},
 	marker::PhantomData,
 };
+
+use crate::tranches::EpochExecutionTranches;
 
 #[derive(Clone, Encode, Decode, Eq, PartialEq, RuntimeDebug, TypeInfo)]
 pub struct ReserveDetails<Balance> {
