+++ libs/types/src/xcm.rs
@@ -1,4 +1,4 @@
-// Copyright 2022 Centrifuge Foundation (centrifuge.io).
+// Copyright 2021 Centrifuge Foundation (centrifuge.io).
 //
 // This file is part of the Centrifuge chain project.
 // Centrifuge is free software: you can redistribute it and/or modify
@@ -10,7 +10,23 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
 
-use cfg_primitives::types::Balance;
+use cfg_primitives::{Balance, Moment, PoolId, TrancheId, TrancheWeight};
+use cfg_traits::TrancheCurrency as TrancheCurrencyT;
+use codec::{Decode, Encode, MaxEncodedLen};
+use frame_support::{
+	dispatch::DispatchResult, ensure, sp_runtime::ArithmeticError, traits::Get, Blake2_128,
+	BoundedVec, Parameter, RuntimeDebug, StorageHasher,
+};
+use orml_traits::asset_registry::AssetMetadata;
+use polkadot_parachain::primitives::Id as ParachainId;
+use rev_slice::{RevSlice, SliceExt};
+use scale_info::TypeInfo;
+use serde::{Deserialize, Serialize};
+use sp_arithmetic::traits::{checked_pow, BaseArithmetic, Unsigned};
+use sp_runtime::{
+	traits::{ConstU32, Member, One, Zero},
+	DispatchError, FixedPointNumber, FixedPointOperand, Perquintill, WeakBoundedVec,
+};
 
 #[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
 #[derive(
