+++ pallets/interest-accrual/src/lib.rs
@@ -1,4 +1,4 @@
-// Copyright 2022 Centrifuge Foundation (centrifuge.io).
+// Copyright 2021 Centrifuge Foundation (centrifuge.io).
 // This file is part of Centrifuge Chain project.
 
 // Centrifuge is free software: you can redistribute it and/or modify
